extract Popup and PopupType into own module
