+++ app/src/lib/dispatcher/dispatcher.ts
@@ -12,8 +12,6 @@ import {
 import { DiffSelection, ImageDiffType } from '../../models/diff'
 import {
   RepositorySectionTab,
-  Popup,
-  PopupType,
   Foldout,
   FoldoutType,
   CompareAction,
@@ -47,7 +45,7 @@ import {
 } from '../../lib/oauth'
 import { installCLI } from '../../ui/lib/install-cli'
 import { setGenericUsername, setGenericPassword } from '../generic-git-auth'
-import { RetryAction, RetryActionType } from '../retry-actions'
+import { RetryAction, RetryActionType } from '../../models/retry-actions'
 import { Shell } from '../shells'
 import { CloneRepositoryTab } from '../../models/clone-repository-tab'
 import { validatedRepositoryPath } from '../../lib/stores/helpers/validated-repository-path'
@@ -60,6 +58,7 @@ import { isGitRepository } from '../git'
 import { ApplicationTheme } from '../../ui/lib/application-theme'
 import { TipState } from '../../models/tip'
 import { RepositoryStateCache } from '../stores/repository-state-cache'
+import { Popup, PopupType } from '../../models/popup'
 
 /**
  * An error handler function.
