+++ app/src/lib/dispatcher/error-handlers.ts
@@ -8,7 +8,7 @@ import { ErrorWithMetadata } from '../error-with-metadata'
 import { ExternalEditorError } from '../editors/shared'
 import { AuthenticationErrors } from '../git/authentication'
 import { Repository } from '../../models/repository'
-import { PopupType } from '../../lib/app-state'
+import { PopupType } from '../../models/popup'
 import { ShellError } from '../shells'
 import { UpstreamAlreadyExistsError } from '../stores/upstream-already-exists-error'
 import { FetchType } from '../../models/fetch'
