+++ app/src/models/popup.ts
@@ -1,0 +1,122 @@
+import { Repository } from './repository'
+import { PullRequest } from './pull-request'
+import { Branch } from './branch'
+import { ReleaseSummary } from './release-notes'
+import { IRemote } from './remote'
+import { RetryAction } from './retry-actions'
+import { WorkingDirectoryFileChange } from './status'
+import { PreferencesTab } from './preferences'
+
+export enum PopupType {
+  RenameBranch = 1,
+  DeleteBranch,
+  ConfirmDiscardChanges,
+  Preferences,
+  MergeBranch,
+  RepositorySettings,
+  AddRepository,
+  CreateRepository,
+  CloneRepository,
+  CreateBranch,
+  SignIn,
+  About,
+  InstallGit,
+  PublishRepository,
+  Acknowledgements,
+  UntrustedCertificate,
+  RemoveRepository,
+  TermsAndConditions,
+  PushBranchCommits,
+  CLIInstalled,
+  GenericGitAuthentication,
+  ExternalEditorFailed,
+  OpenShellFailed,
+  InitializeLFS,
+  LFSAttributeMismatch,
+  UpstreamAlreadyExists,
+  ReleaseNotes,
+  DeletePullRequest,
+  MergeConflicts,
+}
+
+export type Popup =
+  | { type: PopupType.RenameBranch; repository: Repository; branch: Branch }
+  | {
+      type: PopupType.DeleteBranch
+      repository: Repository
+      branch: Branch
+      existsOnRemote: boolean
+    }
+  | {
+      type: PopupType.ConfirmDiscardChanges
+      repository: Repository
+      files: ReadonlyArray<WorkingDirectoryFileChange>
+      showDiscardChangesSetting?: boolean
+    }
+  | { type: PopupType.Preferences; initialSelectedTab?: PreferencesTab }
+  | {
+      type: PopupType.MergeBranch
+      repository: Repository
+      branch?: Branch
+    }
+  | { type: PopupType.RepositorySettings; repository: Repository }
+  | { type: PopupType.AddRepository; path?: string }
+  | { type: PopupType.CreateRepository; path?: string }
+  | {
+      type: PopupType.CloneRepository
+      initialURL: string | null
+    }
+  | {
+      type: PopupType.CreateBranch
+      repository: Repository
+      initialName?: string
+    }
+  | { type: PopupType.SignIn }
+  | { type: PopupType.About }
+  | { type: PopupType.InstallGit; path: string }
+  | { type: PopupType.PublishRepository; repository: Repository }
+  | { type: PopupType.Acknowledgements }
+  | {
+      type: PopupType.UntrustedCertificate
+      certificate: Electron.Certificate
+      url: string
+    }
+  | { type: PopupType.RemoveRepository; repository: Repository }
+  | { type: PopupType.TermsAndConditions }
+  | {
+      type: PopupType.PushBranchCommits
+      repository: Repository
+      branch: Branch
+      unPushedCommits?: number
+    }
+  | { type: PopupType.CLIInstalled }
+  | {
+      type: PopupType.GenericGitAuthentication
+      hostname: string
+      retryAction: RetryAction
+    }
+  | {
+      type: PopupType.ExternalEditorFailed
+      message: string
+      suggestAtom?: boolean
+      openPreferences?: boolean
+    }
+  | { type: PopupType.OpenShellFailed; message: string }
+  | { type: PopupType.InitializeLFS; repositories: ReadonlyArray<Repository> }
+  | { type: PopupType.LFSAttributeMismatch }
+  | {
+      type: PopupType.UpstreamAlreadyExists
+      repository: Repository
+      existingRemote: IRemote
+    }
+  | {
+      type: PopupType.ReleaseNotes
+      newRelease: ReleaseSummary
+    }
+  | {
+      type: PopupType.DeletePullRequest
+      repository: Repository
+      branch: Branch
+      pullRequest: PullRequest
+    }
+  | { type: PopupType.MergeConflicts; repository: Repository }
