+++ app/src/ui/add-repository/add-existing-repository.tsx
@@ -11,7 +11,7 @@ import { Row } from '../lib/row'
 import { Dialog, DialogContent, DialogFooter } from '../dialog'
 import { Octicon, OcticonSymbol } from '../octicons'
 import { LinkButton } from '../lib/link-button'
-import { PopupType } from '../../lib/app-state'
+import { PopupType } from '../../models/popup'
 import * as Path from 'path'
 
 import untildify = require('untildify')
