+++ app/src/ui/generic-git-auth/generic-git-auth.tsx
@@ -6,7 +6,7 @@ import { Button } from '../lib/button'
 import { ButtonGroup } from '../lib/button-group'
 import { Dialog, DialogContent, DialogFooter } from '../dialog'
 import { Monospaced } from '../lib/monospaced'
-import { RetryAction } from '../../lib/retry-actions'
+import { RetryAction } from '../../models/retry-actions'
 
 interface IGenericGitAuthenticationProps {
   /** The hostname with which the user tried to authenticate. */
