+++ app/src/ui/repository-settings/repository-settings.tsx
@@ -5,7 +5,7 @@ import { GitIgnore } from './git-ignore'
 import { assertNever } from '../../lib/fatal-error'
 import { IRemote } from '../../models/remote'
 import { Dispatcher } from '../../lib/dispatcher'
-import { PopupType } from '../../lib/app-state'
+import { PopupType } from '../../models/popup'
 import { Repository } from '../../models/repository'
 import { Button } from '../lib/button'
 import { ButtonGroup } from '../lib/button-group'
